bitbake: parse/ast/data_smart: Add parsing flag to getVar/setVar
When parsing we find problems if we clear prepends/appends when
setting variables during the initial parsing phases. Later, we actively
want to do this (in what would be post finalisation previously).

To handle this, pass a parsing flag to the operations to control
the correct behaviour for the context.

(Bitbake rev: 74d29d908fb79b1a5f748be09648d255e5acd4c7)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>