alsa-plugins, pulseaudio: depend on speexdsp, not speex
Speexdsp was split off from speex, and alsa-plugins and pulseaudio
only depend on the speexdsp library.

alsa-plugins needs also a patch, because it has so far relied on
speex_types.h, which was renamed to speexdsp_types.h when speexdsp was
split off, to not collide with the file that is included in speex.

(From OE-Core rev: 34d62a3e5bdd4f2b0a9373f9d8f9ada8b4e1a791)

Signed-off-by: Tanu Kaskinen <tanu.kaskinen@linux.intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>