oeqa/sstatetests: Add NATIVELSB sstate signature equivalence test
The sstate checksums should be independent of whichever NATIVELSBSTRING is
detected. Add an automated QA test which tests this using bitbake -S.

To make this possible, we need to be able to override the value of
NATIVELSBSTRING so make a small change to allow this.

(From OE-Core rev: b180a5ed864a4e8affe20534c1ba20c9ff8aa016)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>